refactor: address daily quality findings for 2026-08-17
Extract terminal pane drag-preview capture helpers, prepareSavedCenterLayout
for multi-pane restore, and buildTabToPaneIdMap. Archive score-76 report.
